+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 
 ### Fixed
 - Avoid unnecessary switch to `MainConnection` while setting `read-only` mode
+- Statement Behavior changes don't affect DualConnection's state
 
 ## [0.1.24] - 2021-02-16